# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroTwo Bot API
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns overall bot statistics
 * @summary Get bot statistics
 */
export const GetBotStatsResponse = zod.object({
  guildCount: zod.number(),
  userCount: zod.number(),
  commandsExecuted: zod.number(),
  uptime: zod.number(),
  ping: zod.number(),
  botName: zod.string(),
  botAvatar: zod.string().nullish(),
  version: zod.string(),
});

/**
 * Returns recent command executions
 * @summary Get recent bot activity
 */
export const GetBotActivityResponseItem = zod.object({
  id: zod.number(),
  command: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  guildId: zod.string(),
  guildName: zod.string(),
  executedAt: zod.coerce.date(),
  success: zod.boolean(),
});
export const GetBotActivityResponse = zod.array(GetBotActivityResponseItem);

/**
 * @summary List all guilds
 */
export const ListGuildsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  memberCount: zod.number(),
  iconUrl: zod.string().nullish(),
  joinedAt: zod.coerce.date(),
});
export const ListGuildsResponse = zod.array(ListGuildsResponseItem);

/**
 * @summary Get command usage statistics
 */
export const GetCommandStatsResponseItem = zod.object({
  command: zod.string(),
  count: zod.number(),
  lastUsed: zod.coerce.date(),
});
export const GetCommandStatsResponse = zod.array(GetCommandStatsResponseItem);

/**
 * @summary List all warnings
 */
export const ListWarnsQueryParams = zod.object({
  guildId: zod.coerce.string().optional(),
  userId: zod.coerce.string().optional(),
});

export const ListWarnsResponseItem = zod.object({
  id: zod.number(),
  guildId: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  moderatorId: zod.string(),
  moderatorName: zod.string(),
  reason: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListWarnsResponse = zod.array(ListWarnsResponseItem);

/**
 * @summary Create a warning
 */
export const CreateWarnBody = zod.object({
  guildId: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  moderatorId: zod.string(),
  moderatorName: zod.string(),
  reason: zod.string(),
});

/**
 * @summary Delete a warning
 */
export const DeleteWarnParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteWarnResponse = zod.object({
  id: zod.number(),
  guildId: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  moderatorId: zod.string(),
  moderatorName: zod.string(),
  reason: zod.string(),
  createdAt: zod.coerce.date(),
});